+++ src/worker/core/realRosters/getLeague.ts
@@ -333,7 +333,8 @@ const getLeague = async (options: GetLeagueOptions) => {
 			players.push(...freeAgents2);
 		}
 
-		const seenSlugs = new Set();
+		// Find draft prospects, which can't include any active players
+		const seenSlugs = new Set(players.map(p => p.srID));
 		const draftProspects = orderBy(basketball.ratings, ["slug", "season"])
 			.filter(ratings => {
 				// Only keep rookie seasons
@@ -341,13 +342,45 @@ const getLeague = async (options: GetLeagueOptions) => {
 				seenSlugs.add(ratings.slug);
 				return !seen;
 			})
-			.filter(ratings => ratings.season > options.season)
+			.filter(
+				ratings => ratings.season > options.season || options.randomDebuts,
+			)
 			.map(ratings =>
 				formatPlayer(ratings, options.season, initialTeams, {
 					draftProspect: true,
 					randomDebuts: options.randomDebuts,
 				}),
 			);
+
+		if (options.randomDebuts) {
+			// For players in past draft classes, automatically move to future draft classes, so they will be correctly randomized below
+			const targetDraftClassSize = 70;
+			const draftProspectsToMove = draftProspects.filter(
+				p => p.draft.year < options.season,
+			);
+			let draftYear = options.season;
+			let draftClassSize = draftProspects.filter(
+				p => p.draft.year === draftYear,
+			).length;
+			while (draftProspectsToMove.length > 0) {
+				while (draftClassSize >= targetDraftClassSize) {
+					draftYear += 1;
+					draftClassSize = draftProspects.filter(
+						p => p.draft.year === draftYear,
+					).length;
+				}
+
+				const p = draftProspectsToMove.pop();
+				if (p) {
+					p.tid = PLAYER.UNDRAFTED;
+
+					const diff = draftYear - p.draft.year;
+					p.draft.year += diff;
+					p.born.year += diff;
+					draftClassSize += 1;
+				}
+			}
+		}
 
 		players.push(...draftProspects);